Rework ros nodes + core modules for new project + remove simulation files

+++ av_navigation/core/include/planner_module/point_cloud.hpp
@@ -1,10 +1,11 @@
 #pragma once
 
 #include <vector>
-#include "Eigen/Dense"
+#include "eigen3/Eigen/Dense"
 
 
 namespace NAVIGATION_CORE {
+
 
 struct PointXYZ {
     PointXYZ() : x(0.0f), y(0.0f), z(0.0f) {}
@@ -14,8 +15,10 @@ struct PointXYZ {
     float z = 0.0f;
 };
 
+
 inline Eigen::Vector3f toEigen(const PointXYZ& p) {return Eigen::Vector3f(p.x, p.y, p.z);}
 inline PointXYZ toPointXYZ(const Eigen::Vector3f& p) {return PointXYZ(p.x(), p.y(), p.z());}
+
 
 struct PointXYZI {
     PointXYZI() : x(0.0f), y(0.0f), z(0.0f), i(0.0f) {}
@@ -26,8 +29,10 @@ struct PointXYZI {
     float i = 0.0f;
 };
 
+
 inline Eigen::Vector3f toEigen(const PointXYZI& p) {return Eigen::Vector3f(p.x, p.y, p.z);}
 inline PointXYZI toPointXYZI(const Eigen::Vector3f& p, float age = 0.0f) {return PointXYZI(p.x(), p.y(), p.z(), age);}
+
 
 template <typename PointType>
 class PointCloud {
@@ -36,7 +41,7 @@ public:
     PointCloud() = default;
     ~PointCloud() = default;
 
-    using iterator       = typename std::vector<PointType>::iterator;
+    using iterator = typename std::vector<PointType>::iterator;
     using const_iterator = typename std::vector<PointType>::const_iterator;
 
     iterator begin() {return this->cloud_.begin();}
@@ -51,9 +56,10 @@ public:
     void clear() {this->cloud_.clear();}
     size_t size() const {return this->cloud_.size();}
 
-    PointType& operator[](int index) {return this->cloud_[index];}
+    PointType& operator[](size_t index) {return this->cloud_[index];}
 
     std::vector<PointType> cloud_ = {};
 };
+
 
 } // namespace NAVIGATION_CORE
